+++ internal/provider/resource_tfe_oauth_client.go
@@ -22,6 +22,7 @@ func resourceTFEOAuthClient() *schema.Resource {
 		Create: resourceTFEOAuthClientCreate,
 		Read:   resourceTFEOAuthClientRead,
 		Delete: resourceTFEOAuthClientDelete,
+		Update: resourceTFEOAuthClientUpdate,
 
 		CustomizeDiff: customizeDiffIfProviderDefaultOrganizationChanged,
 
@@ -112,6 +113,12 @@ func resourceTFEOAuthClient() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+
+			"organization_scoped": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Default:  true,
+			},
 		},
 	}
 }
@@ -139,12 +146,13 @@ func resourceTFEOAuthClientCreate(d *schema.ResourceData, meta interface{}) erro
 	// The tfe.OAuthClientCreateOptions has omitempty for these values, so if it
 	// is empty, then it will be ignored in the create request
 	options := tfe.OAuthClientCreateOptions{
-		Name:            tfe.String(name),
-		APIURL:          tfe.String(d.Get("api_url").(string)),
-		HTTPURL:         tfe.String(d.Get("http_url").(string)),
-		OAuthToken:      tfe.String(d.Get("oauth_token").(string)),
-		Key:             tfe.String(key),
-		ServiceProvider: tfe.ServiceProvider(serviceProvider),
+		Name:               tfe.String(name),
+		APIURL:             tfe.String(d.Get("api_url").(string)),
+		HTTPURL:            tfe.String(d.Get("http_url").(string)),
+		OAuthToken:         tfe.String(d.Get("oauth_token").(string)),
+		Key:                tfe.String(key),
+		ServiceProvider:    tfe.ServiceProvider(serviceProvider),
+		OrganizationScoped: tfe.Bool(d.Get("organization_scoped").(bool)),
 	}
 
 	if serviceProvider == tfe.ServiceProviderAzureDevOpsServer {
@@ -189,6 +197,7 @@ func resourceTFEOAuthClientRead(d *schema.ResourceData, meta interface{}) error 
 	d.Set("api_url", oc.APIURL)
 	d.Set("http_url", oc.HTTPURL)
 	d.Set("service_provider", string(oc.ServiceProvider))
+	d.Set("organization_scoped", oc.OrganizationScoped)
 
 	switch len(oc.OAuthTokens) {
 	case 0:
@@ -215,4 +224,21 @@ func resourceTFEOAuthClientDelete(d *schema.ResourceData, meta interface{}) erro
 	}
 
 	return nil
+}
+
+func resourceTFEOAuthClientUpdate(d *schema.ResourceData, meta interface{}) error {
+	config := meta.(ConfiguredClient)
+
+	// Create a new options struct.
+	options := tfe.OAuthClientUpdateOptions{
+		OrganizationScoped: tfe.Bool(d.Get("organization_scoped").(bool)),
+	}
+
+	log.Printf("[DEBUG] Update OAuth client %s", d.Id())
+	_, err := config.Client.OAuthClients.Update(ctx, d.Id(), options)
+	if err != nil {
+		return fmt.Errorf("Error updating OAuth client %s: %w", d.Id(), err)
+	}
+
+	return resourceTFEOAuthClientRead(d, meta)
 }