+++ internal/provider/resource_tfe_sentinel_version_test.go
@@ -255,8 +255,8 @@ func genSentinelSha(t *testing.T, secret, data string) string {
 }
 
 // genSafeRandomSentinelVersion returns a random version number of the form
-// `0.0.<RANDOM>`, which TFC won't ever select as the latest available
-// Sentinel. (At the time of writing, a fresh TFC instance will include
+// `0.0.<RANDOM>`, which HCP Terraform won't ever select as the latest available
+// Sentinel. (At the time of writing, a fresh HCP Terraform instance will include
 // official Sentinels 0.22.1 and higher.) This is necessary because newly created
 // workspaces default to the latest available version, and there's nothing
 // preventing unrelated processes from creating workspaces during these tests.